Make RecvMsg useful by addding an optional MinLength. Generally replace uses of Send/RecvPartial with Send/RecvMsg.
Simplify use of ShowDelay in StreamServer.

Change various ULONG types to size_t in SSLClient and Utilities.

+++ Common/Include/Utilities.h
@@ -10,8 +10,8 @@
 
 std::wstring string_format(const WCHAR* pszFormat, ...);
 std::wstring WinErrorMsg(int nErrorCode);
-void PrintHexDump(DWORD length, const void * const buf);
-void PrintHexDump(DWORD length, const void * const buf, const bool verbose);
+void PrintHexDump(size_t length, const void * const buf);
+void PrintHexDump(size_t length, const void * const buf, const bool verbose);
 void SetThreadName(char* threadName);
 void SetThreadName(char* threadName, DWORD dwThreadID);
 void DebugMsg(const WCHAR* pszFormat, ...);
